feat: refactor LogViewButton and DetailLogView
