+++ src/components/Deploy/DeployLogs/index.tsx
@@ -1,3 +1,4 @@
 export * from './DeploysInfos'
 export * from './DeployLogs'
 export * from './OperationsList'
+export * from './DetailLogView'
